Create synthetic note class and move client document note and outbound call notes to messages tab
